Issue#10 Refactor Console UI
- Split pages into different files.
- Remove obsolete console.

Fixes #10
