+++ src/main.cpp
@@ -1,53 +1,17 @@
 #include <iostream>
 #include <string>
-#include "console.cpp"
+
+#include "productservice.h"
+#include "transactionservice.h"
+
+#include "ui/console/mainmenu.h"
 
 using namespace std;
 
-class GroceriesInsight {
-public:
-	GroceriesInsight() {
-		console;
-	}
-	
-	void run() {
-		ProductService::load();
-		TransactionService::load();
-		
-		int option;
-		do {
-			console.showMainMenu();
-			
-			option = Console::readInteger();
-			switch (option) {
-				case 1:
-					console.showAddProductPage();
-					break;
-				case 2:
-					console.showSearchProductPage();
-					break;
-				case 3:
-					console.showAddTransactionPage();
-					break;
-				case 4:
-					console.showAllTransactionsPage();
-					break;
-				case 5:
-					console.showProductHistoryPage();
-					break;
-				case 6:
-					console.showExpenditureHistoryPage();
-					break;
-				default:
-					break;
-			}
-		} while (option != -1);
-	}
-private:
-	Console console;
-};
-
 int main(int argc, const char* argv[]) {
-	GroceriesInsight application;
-	application.run();
+	ProductService::load();
+	TransactionService::load();
+	
+	MainMenu menu;
+	menu.show();
 }
